+++ Clippit/Word/WmlToHtmlConverter.cs
@@ -6,7 +6,7 @@ using System.Globalization;
 using System.Text;
 using System.Xml.Linq;
 using DocumentFormat.OpenXml.Packaging;
-using Image = SixLabors.ImageSharp.Image;
+using SkiaSharp;
 
 // 200e lrm - LTR
 // 200f rlm - RTL
@@ -106,7 +106,7 @@ namespace Clippit.Word
     [SuppressMessage("ReSharper", "UnusedMember.Global")]
     public class ImageInfo
     {
-        public Image Image;
+        public SKBitmap Image;
         public XAttribute ImgStyleAttribute;
         public string ContentType;
         public XElement DrawingElement;
@@ -3550,7 +3550,9 @@ namespace Clippit.Word
                 return null;
 
             using var partStream = imagePart.GetStream();
-            using var image = Image.Load(partStream);
+            using var image = SKBitmap.Decode(partStream);
+            if (image == null)
+                return null;
             if (extentCx != null && extentCy != null)
             {
                 var imageInfo = new ImageInfo()
@@ -3625,7 +3627,10 @@ namespace Clippit.Word
                 using var partStream = imagePart.GetStream();
                 try
                 {
-                    using var bitmap = Image.Load(partStream);
+                    using var bitmap = SKBitmap.Decode(partStream);
+                    if (bitmap == null)
+                        return null;
+
                     var imageInfo = new ImageInfo
                     {
                         Image = bitmap,